+++ R/var_label.R
@@ -1,11 +1,14 @@
 #' Get / Set a variable label
 #'
 #' @param x An object.
-#' @param value A character string or \code{NULL} to remove the label. For data frames,
-#'   it could also be a named list.
+#' @param value A character string or \code{NULL} to remove the label.
+#'  For data frames, it could also be a named list or a character vector
+#'  of same length as the number of columns in \code{x}.
 #' @details
 #'   For data frames, if \code{value} is a named list, only elements whose name will
-#'   match a column of the data frame will be taken into account.
+#'   match a column of the data frame will be taken into account. If \code{value}
+#'   is a character vector, labels should in the same order as the columns of the
+#'   data.frame.
 #' @examples
 #' var_label(iris$Sepal.Length)
 #' var_label(iris$Sepal.Length) <- 'Length of the sepal'
@@ -14,6 +17,17 @@
 #' }
 #' # To remove a variable label
 #' var_label(iris$Sepal.Length) <- NULL
+#' # To change several variable labels at once
+#' var_label(iris) <- c(
+#'   "sepal length", "sepal width", "petal length",
+#'   "petal width", "species"
+#')
+#' var_label(iris)
+#' var_label(iris) <- list(
+#'   Petal.Width = "width of the petal",
+#'   Petal.Length = "length of the petal"
+#' )
+#' var_label(iris)
 #' @export
 var_label <- function(x) {
   UseMethod("var_label")
@@ -48,11 +62,15 @@ var_label.data.frame <- function(x) {
 #' @export
 `var_label<-.data.frame` <- function(x, value) {
   if ((!is.character(value) & !is.null(value)) & !is.list(value) |
-    (is.character(value) & length(value) > 1))
-    stop("`value` should be a single character string, NULL or a named list",
+    (is.character(value) & length(value) > 1 & length(value) != ncol(x)))
+    stop("`value` should be a named list, NULL, a single character string or a character vector of same length than the number of columns in `x`",
       call. = FALSE, domain = "R-labelled")
-  if (is.character(value)) {
+  if (is.character(value) & length(value) == 1) {
     value <- as.list(rep(value, ncol(x)))
+    names(value) <- names(x)
+  }
+  if (is.character(value) & length(value) == ncol(x)) {
+    value <- as.list(value)
     names(value) <- names(x)
   }
   if (is.null(value)) {